CB-3781 use build in codemirror portal
